Many smaller changes
Export fixed for collision models.
Markers were not correctly exported unless the robot was in rest
position (even then there were problems). This could be solved using
the PoseBone elements of the armature, which have the required matrix
available. This could be of interest for exporting the visual scene
when the rest position is left.
Collision meshes were also added.